// temper/shared-addon-libraries-lib-debug-logger/src/types.ts
export type LogEntry = [
  number,
  string,
  number,
  string,
  string,
  string | readonly string[],
  (string | readonly string[])?,
  number?,
]

export interface Settings {
  version: number
  logTraces: boolean
  minLogLevel: string
  loadScreenStartTime: number
  [key: string]: unknown
}

export interface LoggerConfig {
  tag: string
  minLevelOverride?: string
  logTracesOverride?: boolean
}